Calculations for same proceedings with relevant orders
This commit modifies the calculations to ignore relevant orders.
This prevents a relevant order to drag-through in case it's the longest sentence in the group.
Removing the relevant order when it's not the longest won't affect the desired effect either as some other sentence will prevail.

It won't remove relevant orders when all sentences are relevant orders - it keeps the logic as it was before
It will keep the same logic as before when there are no relevant orders.

---

We need clarification to workout what is the correct behaviour when there are relevant orders,
Our current briefing on this matter is:

"if a conviction contains multiple sentences where one is a relevant order, the relevant order is not taken into account for drag-through purposes"

Although this does not clarify a scenario where there are 3 convictions and 2 convictions are relevant orders and are also the longest of the 3.

I've played the following scenario in Unlock calculator and they gave me the longest conviction (1st conviction)

They all started in 1/1/2000 (same proceedings)
1st has a length of 48 months (longet relevant order)
2nd has a length of 36 months (second relevant order)
3rd is a motoring fine, should have expired in 6 months without endorsement

I've created another card to revisit this calculations again, this card can be found here: https://trello.com/c/sZ7qBDwe/105-dc-double-check-that-the-calculations-for-multiple-relevant-orders-in-same-proceedings-are-correct

Story: https://trello.com/c/ImE62qTS/95-dc-multiples-same-proceedings-scenarios-for-calculations-testing

Fix