Show conviction date step in multiples MVP
Ask this new date, but only if we are running the "multiples" MVP (so it will not show for simples or in production).

Unfortunately to do this we have to add some conditional code and propagate the `multiples_enabled` boolean to the decision trees as these do not have access to the controller/session and can't read the cookie.