Validate `known_date` against the `conviction_date`
Ticket: https://trello.com/c/1eRx0JeU

Little follow-up to PR #435 to add a validation to the form so the `known_date` can't be *before* the `conviction_date` as that would not make sense and could invalidate the calculations without the user even noticing.

If `conviction_date` or `known_date` are `nil` then we don't perform this validation (currently for "simples" the `conviction_date` is always nil so we must maintain backwards compatibility).