Add Scenario 2 example from Nacro Organisation
Alex was then convicted of battery on 10/06/11. He was sentenced to:
• A six-month custodial sentence (would become ‘spent’ 10/12/13)
• Restraining order until further order (will be ‘unspent’ until further notice)
• A fine of £150.00 (would become ‘spent’ 10/06/12)

In this case the conviction for battery will remain ‘unspent’ due to the
restraining order until further notice having the longest rehabilitation period.
Therefore, it will be disclosed on a basic DBS certificate indefinitely.

Beware this was adapted, the whole scenario is compose of two separate
convictions, of which, this example is the second conviction.