More tweaks to the MultipleOffensesCalculator
Ticket: https://trello.com/c/PcAlcOaz

This work relates to the edge case where under certain circunstances a never spent, or indefinite conviction could take over the other convictions even if there was no overlapping between them.

In doing this I realised the loop can be simplified further, if we trick the special variants into behave like a Date in terms of ranges.

What this means is:
- A `ResultsVariant::NEVER_SPENT` is like an infinity date, it has no end.
- A `ResultsVariant::INDEFINITE` is a very far in the future date, but is not infinity (this difference is important as one takes precedence over the other).

There might be further tweaks/fixes to do and also it is still pending the work for relevant orders that _might_ affect the loop or the date we need to use on it. But that will come in separate PR(s).