Adds relevant orders to BaseMultiplesCalculator
Follow up on #422 where the method .relevant_order? was added to DisclosureCheck.

We need this capabilities of ignoring relevant orders when comparing two convictions,
relevant orders only affect their own Conviction without interfering with other Convictions.

By discarding the relevant orders, we will be able to loop through several
convictions without compromising their spent date with the relevant order spent
date.

I've added this method in BaseMultiplesCalculator as I think we may need the
method in both SameProceedings and SeparateProceedings so that it can be used
in MultipleOffensesCalculator.

This addition is a non-breaking change to current behaviour.

Story: https://trello.com/c/7scNWWCF/