Fixes overlapping convictions spent_dates
We almost got it correct before.

To solve the issue of the scenario represented in the graph I
had to compare the range of dates without the relevant order.

But first one needs to confirm if a conviction only holds a relevant order,
if so, skip the range comparison.

If there's a date that does not belong to a relevant order,
then we compare the overlapping, but never with relevant order dates.

My previous mistake was to only consider 1 proceeding, this now considers both proceedings

Story: https://trello.com/c/XJEnGNj6/109-ignore-relevant-order-sentences-in-other-separate-overlapping-convictions