Add approximate date copy to results page
Ticket: https://trello.com/c/L2pqJMHo

First part of the work, to add an additional line of copy when we detect there is at least one approximate date in the check.

The other line of copy, shown always, has been tweaked according to content designer.

In a follow-up PR the copy will be implemented in the hint texts.