Amend hint text to clarify approximate dates
Ticket: https://trello.com/c/L2pqJMHo

Follow-up to PR #449.

We want to introduce some extra copy in all the dates hint text. However in doing this I realised we had many different hint texts depending if it was a caution/conviction and the type of the offence.

In order to make this more easy to maintain so we only have 1 hint text for all dates and it appears in the same place always, I've moved out from the hint text and into the `lead` paragraph some of the copy previously had in the hint text itself.