Expose a way to share/return to the results page
This, for now, it is only for development/test purposes so it is hidden behind the "developer tools" footer menu.

Maybe in the future this is a feature we can expand and offer to end users so they can also share their results with other people.

The mechanism is simple. We use the ID of the `disclosure_report` to get the checks inside, and we pick one and assign their ID to the session, redirecting back to the results page to render the results by running the calculations again from the database.