Copy changes for youth cautions in results page
Ticket: https://trello.com/c/eRalDg7Z

Legislation changes were made that relate to Disclosure Checker on 28 November 2020. These were brought to the team's attention on 11 January.

Youth cautions, reprimands and warnings will no longer be automatically disclosed to employers on a Standard or Enhanced Disclosure and Barring Service (DBS) certificate.

Copy changes in the results page when we detect it is a youth caution. No changes for adult cautions (or convictions).