Improve calculator with a memory
This is supposed to "fix" yet another edge case, but using a very basic "memory" where we store the spent date after any drag-though so in the following iteration of the loop, we use it if present and keep dragging the date if needed (and store back again in the memory).

So far I've made the minimum amount of changes but there are other approaches.