Improve documentation to explain multiple convictions
I've added references to the graphs on tests that I thought it would be beneficial,
I did not want to add them on all as it could cause potential misunderstandings.

I've also added references to the README.md so that newcomers can see it and get familiar with
the most important part of this service - the calculations.

Ticket: https://trello.com/c/kT50EDZ7