Update calculations to use drag through reference
Now that we know which relevant orders drag through we can use that flag
to perform calculations correctly.

Only scenario that was flagged by Legal was Scenario 1,
that was the only scenario that failed after using drag through,
this means that we are aligned with Legal.

Story: https://trello.com/c/R9VWuOAu/