Check Your Answers Page
I've done a soft separation from CheckYourAnswers, which was being shared
between the results page and the Check Your Answers page.

Having MultipleResultsPresenter enable us to separate both pages, althought for now,
the only thing that is different is the `results_page` attribute that is passed
down the chain of existing classes.

I don't want to commit to fully separate both right now, it's not clear to me
at this moment what we will need once we start working on the results page, so
I rather have a soft transition here.

-----

I've also requested a decision to show number before the word "Conviction" (or "Caution"),
as this felt misleading to have the number after the word.

Prototype suggested

Conviction 1
Conviction 2

The problem emerged when a caution is added after a conviction, like so:

Conviction 1
Conviction 2
Caution 3

It could make the user think that they have a 3rd caution when they only inserted 1 caution.
So we have agreed to have the number before as if it was a numbered bullet point list.

1. Conviction
2. Conviction
3. Caution

-----

Header logic for sentence numeration

When a conviction has 1 sentence, do not show the header ("Sentence 1")
When a conviction has more than 1 sentence, show the header with numeration

Story: https://trello.com/c/eLQkrUGd