Refactor code to fit the steps framework better
A bit of refactor so the flow fits better the step framework we have.

Note: the `back` link in the check kind, when adding a new caution or conviction, no matter what we set in the navigation stack is hardcoded to the GDS page. But for now, let's leave it, will figure out something later.