Implement Scenario 8 (by Legal)
We have learned about another edge case around relevant orders, so far our
understanding was that a relevant order would be ignored when comparing
convictions.

Although, when a relevant order is the sentence with the longest period within
it's own conviction AND is the longest when comparing with another conviction,
it remains the longest in it's own conviction, and does not affect the external
conviction.

For a better understanding, I've created and added a new graph, this can be seen in:
docs/results/08_relevant_order_3.png

Story: https://trello.com/c/ycbp7CgU