Add new catch-all requirement order option
Ticket: https://trello.com/c/7SK4PLNz

This is the 2nd part of the work requested in the above ticket.

A new catch-all radio option in the "Community, prevention or reparation order" (note this bucket will be renamed once the PR #485 is merged) has been added, named "Any other order with a requirement".

At the moment this is a relevant order, with no drag through, and the calculation rules are the same as in the community order conviction.

However we need confirmation if these rules are correct before merging. Other than that the rest of the work is done.